# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EMS API Documentation",
        "description": "Dokumentasi API untuk event management system:<br>Auth: Register, Login, dan Logout<br>Event: daftar event, detail event, pendaftaran user<br>Attendance: check-in, check-out, dan riwayat",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Server utama API EMS"
        }
    ],
    "paths": {
        "/api/v1/attendance/check-in": {
            "post": {
                "tags": [
                    "Attendance"
                ],
                "summary": "Melakukan check-in ke event",
                "description": "User melakukan check-in dengan QR code untuk event tertentu.",
                "operationId": "438f1046261525c5db6b32818df8c7e8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "event_id",
                                    "qr_code_value"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "ID user yang melakukan check-in",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "event_id": {
                                        "description": "ID event yang diikuti",
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "qr_code_value": {
                                        "description": "Nilai QR code unik event",
                                        "type": "string",
                                        "example": "EVT2025-USER01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Check-in berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Checked in"
                                        },
                                        "data": {
                                            "properties": {
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "event_id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "check_in_time": {
                                                    "type": "string",
                                                    "format": "datetime",
                                                    "example": "2025-10-16T09:00:00Z"
                                                },
                                                "qr_code_value": {
                                                    "type": "string",
                                                    "example": "EVT2025-USER01"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/attendance/check-out": {
            "post": {
                "tags": [
                    "Attendance"
                ],
                "summary": "Melakukan check-out dari event",
                "description": "User melakukan check-out dari event yang sudah diikuti.",
                "operationId": "8036eaf758a73a517bff4863728ba79a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "event_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "event_id": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Check-out berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Checked out"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data absensi tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/attendance/history": {
            "get": {
                "tags": [
                    "Attendance"
                ],
                "summary": "Melihat riwayat absensi user",
                "description": "Mengambil semua data absensi (check-in & check-out) milik user yang sedang login.",
                "operationId": "8170da430227d361ae10adb33dad913f",
                "responses": {
                    "200": {
                        "description": "Daftar riwayat absensi user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 10
                                            },
                                            "event_id": {
                                                "type": "integer",
                                                "example": 5
                                            },
                                            "check_in_time": {
                                                "type": "string",
                                                "format": "datetime",
                                                "example": "2025-10-16T09:00:00Z"
                                            },
                                            "check_out_time": {
                                                "type": "string",
                                                "format": "datetime",
                                                "example": "2025-10-16T17:00:00Z"
                                            },
                                            "event": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Laravel Dev Conference 2025"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}